pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]
